chore: Extracted cluster fom dataclasses into their own files; adapted imports
